lang updates, dashboard detail view
