+++ mod-disco/service/flu/lib/modules/dashboard/views/dashboard_detail_view.dart
@@ -1,105 +1,33 @@
-import 'dart:typed_data';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_modular/flutter_modular.dart';
+import 'package:mod_disco/core/config/env_config.dart';
 import 'package:mod_disco/core/core.dart';
+import 'package:mod_disco/core/routes/paths.dart';
 import 'package:mod_disco/modules/dashboard/view_model/dashboard_detail_view_model.dart';
-import 'package:mod_disco/modules/dashboard/view_model/dashboard_view_model.dart';
 import 'package:mod_disco/modules/dashboard/widgets/data_pane/data_pane.dart';
 import 'package:mod_disco/modules/dashboard/widgets/filter_pane.dart';
-// import 'package:mod_disco/modules/org_manager/orgs/data/org_model.dart';
 import 'package:provider_architecture/provider_architecture.dart';
 import 'package:responsive_builder/responsive_builder.dart';
-import 'package:sys_core/sys_core.dart';
-import 'package:sys_share_sys_account_service/rpc/v2/sys_account_models.pb.dart';
 
-class OrgMasterDetailView extends StatelessWidget {
-  final String id;
+class DashboardDetailView extends StatelessWidget {
   final String orgId;
-  final String routePlaceholder;
+  final String projectId;
+  final bool isFullScreen;
 
-  OrgMasterDetailView(
-      {Key key, this.id = '', this.orgId = '', this.routePlaceholder})
-      : super(key: key);
+  DashboardDetailView({
+    Key key,
+    @required this.orgId,
+    @required this.projectId,
+    @required this.isFullScreen,
+  });
 
   @override
   Widget build(BuildContext context) {
     return ViewModelProvider.withConsumer(
-      viewModelBuilder: () => DashboardViewModel(),
-      onModelReady: (DashboardViewModel model) async {
-        if (model.orgs == null || model.orgs.isEmpty) {
-          await model.getPermissions();
-          await model.getInitialAdminOrgs();
-        }
-      },
-      builder: (context, DashboardViewModel model, child) =>
-          GCMasterDetail<Org, Project>(
-        enableSearchBar: true,
-        parentId: orgId,
-        childId: id,
-        items: model.orgs,
-        labelBuilder: (item) => item.name,
-        imageBuilder: (item) => item.logo,
-        routeWithIdPlaceholder: routePlaceholder,
-        detailsBuilder: (context, parentId, childId, isFullScreen) {
-          model.getSelectedProjectAndDetails(parentId, childId);
-          return _getDetailsView(context, parentId, childId, isFullScreen);
-        },
-        noItemsAvailable: Center(
-          child: Text(
-            ModDiscoLocalizations.of(context).translate('noCampaigns'),
-          ),
-        ),
-        noItemsSelected: Center(
-            child: Text(ModDiscoLocalizations.of(context)
-                .translate('noItemsSelected'))),
-        disableBackButtonOnNoItemSelected: false,
-        masterAppBarTitle:
-            Text(ModDiscoLocalizations.of(context).translate('selectCampaign')),
-        isLoadingMoreItems: model.isLoading,
-        fetchNextItems: model.getNextAdminOrgs,
-        hasMoreItems: model.hasMoreItems,
-        searchFunction: model.searchAdminOrgs,
-        resetSearchFunction: model.onResetSearchOrgs,
-        itemChildren: (org) => org.projects,
-        childBuilder: (project, id) => <Widget>[
-          ...[
-            SizedBox(width: 16),
-            CircleAvatar(
-              radius: 20,
-              backgroundImage: MemoryImage(
-                Uint8List.fromList(project.logo),
-              ),
-            ),
-          ],
-          SizedBox(width: 16),
-          //logic taken from ListTile
-          Expanded(
-            child: Text(
-              project.name,
-              textAlign: TextAlign.center,
-              style: Theme.of(context).textTheme.subtitle1.merge(
-                    TextStyle(
-                      color: project.id != id
-                          ? Theme.of(context).textTheme.subtitle1.color
-                          : Theme.of(context).accentColor,
-                    ),
-                  ),
-            ),
-          ),
-          SizedBox(width: 30),
-        ],
-      ),
-    );
-  }
-
-  Widget _getDetailsView(BuildContext context, String parentId, String childId,
-      bool isFullScreen) {
-    return ViewModelProvider.withConsumer(
       viewModelBuilder: () => DashboardDetailViewModel(
-        orgId: parentId,
-        projectId: childId,
+        orgId: orgId,
+        projectId: projectId,
       ),
       onModelReady: (DashboardDetailViewModel model) async =>
           await model.fetchUserDatas(),
@@ -110,15 +38,16 @@ class OrgMasterDetailView extends StatelessWidget {
               appBar: AppBar(
                 // iconTheme: Theme.of(context).iconTheme,
                 automaticallyImplyLeading: isFullScreen,
-                title: Text('Project Dashboard'),
+                title:
+                    Text(ModDiscoLocalizations.of(context).projectDashboard()),
                 // this the mock data
                 actions: <Widget>[
                   IconButton(
-                      tooltip: "Copy Link",
+                      tooltip: ModDiscoLocalizations.of(context).copyLink(),
                       icon: Icon(Icons.link),
                       onPressed: () async {
                         String link =
-                            "${Modular.get<EnvConfig>().url}/#/${Modular.get<DashboardPaths>().dashboardId.replaceFirst("/", "").replaceAll(":orgId", "$parentId").replaceAll(":id", childId)}";
+                            "${Modular.get<EnvConfig>().url}/#/${Modular.get<DashboardPaths>().dashboardId.replaceFirst("/", "").replaceAll(":orgId", "$orgId").replaceAll(":id", projectId)}";
                         await Clipboard.setData(new ClipboardData(text: link));
                       })
                 ],
@@ -161,7 +90,6 @@ class OrgMasterDetailView extends StatelessWidget {
                                 // TODO: roles filter
                                 sizingInfo: sizingInfo,
                               )
-                            // ? Container(child: Text('FilterPane here'))
                             : Offstage(),
                         SizedBox(width: 16),
                         Expanded(
